docs: merge docstrings of TDVP methods
Gather the documentation for all methods of each single TDVP variant
in a single docstrings, so that e.g. `tdvp1!` has just one docstring
instead of a different one for each different method (there's one with
a single MPO for the Hamiltonian, another for a Vector of MPOs, and yet
another one with a ProjMPO). This makes the documentation much more
readable, since there's now a single "point of access" for each TDVP
function, making it easier to tell the difference between the methods.

The names of some variables were also simplified, e.g. using `dt` for
the time step everywhere, and avoiding all non-ASCII characters except
in mathematical formulas within docstrings.