feat: remove `site` from `apply!` for vect. TDVP1
Write a specific version of `apply!` (as well as `measure_localops!`)
for the vectorised TDVP1 algorithm that does not take the site number as
argument.
The measurement routine does not depend on the site, so it's a useless
argument. By promoting `alg` to a (mandatory) positional argument in
`apply!`, multiple dispatch can be used to call the appropriate function
depending on the evolution algorithm used, without resorting to if-else
blocks inside the function body.

Moreover, in the `tdvp1vec!` function, `apply!` was called inside a loop
that cycled over all sites of the state MPS. This is enormously wasteful
since it means that the very same operation is repeated as many times
as the number of sites in the MPS, since in this case `apply!` does not
depend on the particular site. This loop has been removed.