feat: optimise measurement step in `tdvp1!`
Move the measurement step `apply!` in the middle of the evolution
routine, so that the canonicity of the MPS can be leveraged to contract
only the sites in the domain of the operator we want to measure.
This is a great simplification from the full MPS-MPO-MPS contraction
that was there before, but requires that we tear the `tdvp_site_update!`
function apart to place `apply!` in the middle; as a result, until a
more elegant solution that doesn't affect the performance is found, the
`tdvp1!` function doesn't call `tdvp_site_update!` anymore in its body.