feat: optimise measurement step in `jointtdvp1!`
Simplify the measurement step within the `jointtdvp1!` function so that
it is executed only once at the end of the sweep, since---like in the
vectorised TDVP scheme---there are no orthogonality rules we can use to
simplify the contraction anyway. Replace also the full MPS-MPO-MPS
contraction with a simpler contraction where we apply the local operator
directly as an ITensor operator to the MPS, locally on the domain, so that
there's no time lost in contracting with identities the tensors
lying outside of the operator domain.