feat!: optimise measurement on vectorised states
Rewrite the `measure_localops!` function in the vectorised TDVP1 case so
that partial contractions between local identities and single sites from
the state are saved and recycled, between the measurement of an operator
and the next one in the callback.
Previously, a whole MPS-MPS contraction was called for each expectation
value, so this saves a lot of time, especially if there's a lot of
operators that need to be evaluated; even if the operator MPS was
cached so that it didn't need to be recreated at each measurement step,
the contraction took a significant amount of time.

BREAKING CHANGE: it is not necessary anymore to specify the adjoint of
the operator one wants to measure, in the `LocalOperator` constructor,
nor to add a `v` to the operator names (which is added automatically).
This way the operators to be measured can be specified, as
`LocalOperators`, in the exact same way for both `tdvp1!` and
`tdvp1vec!` functions.

The previous contraction method called `inner`, which conjugated the
MPS representing the operator: as a consequence, if the user wanted
to measure a non-Hermitian operator, say `A` for fermions or bosons,
they would have to use `LocalOperator(n => "Adag")` so that the inner
product correctly computed ``⟨A*, ρ⟩ = tr(A ρ)``.
The new contraction method works so that the tensors encoding the
operator are not conjugated, and this manual correction is not needed
anymore.