refactor: measure norm together with observables
Take advantage of the saved partial contraction in the measurement step
by computing the norm (or trace) too in the same function, instead of
doing that separately in the `printoutput` function.
Also output both real and imaginary part of the norm or trace; both are
computed anyway so we might as well print them both.