test: compare TDVP time series by element
Compare the results of the TDVP tests element by element, i.e. by
calling `isapprox` on each element of the expectation value time series
instead of on the whole time series. Overall this seems to be more
correct logically, because a small error which is constant in time is
tolerable, but if the whole `Vector`s of expectation values are compared
against one another this would show up as a bigger error.