feat!: invert ancillary sites in SF formalism
Apply a particle-hole transformation on the ancillary sites in the
superfermion machinery: this allows the generator of the GKSL equation
to be written in a way that conserves the total number of fermions,
enabling a more aggressive QN conservation on the MPS.

BREAKING CHANGE: all initial states of superfermion MPSs and all
operators acting on superfermion MPSs must implement this inversion by
changing the states/operators relative to the ancillary sites
accordingly. Fermion JW string operators are not affected (the JW
transformation can take place after the particle-hole inversion so it
continues to work as before).